ENH: Adding Analysis Configurations to Template files
This will enable template autions to specify analaysis configurations in the XML
without having to know the resulting definition structure.

The following is an example snippet - for more infomation look at data/attribute/attribute_collection/analysisConfigTest.sbt and the test unitAnalysisConfiguration.cxx.

Also fixed an issue with exclusive analyses that required the children's label to match their name.

Also fixed Windows compiler warnings in unitExclusionCategories.cxx and qtFileItem.cxx.

Closes #398 #397

+++ smtk/attribute/testing/cxx/unitExclusionCategories.cxx
@@ -45,7 +45,7 @@ bool testCategories(const attribute::ResourcePtr& attRes, const std::string& pre
     std::cerr << "\t Testing Configuration: " << count;
     for (i = 0; i < n; i++)
     {
-      auto idef = def->itemDefinition(i);
+      auto idef = def->itemDefinition(static_cast<int>(i));
       if (idef->localCategories().passes(test.first) != test.second[i])
       {
         std::cerr << " " << i++;
